refactor(scheduler): address PR #720 reviewer feedback
- Move SystemClock out of mofa-kernel into mofa-foundation/src/scheduler/clock.rs
  (kernel layer must contain only trait definitions, no concrete impls)
- Remove #[async_trait::async_trait] from AgentScheduler; use native async fn
  in trait (Rust 1.75+)
- Rename AgentScheduler::resume_schedule → resume for consistency with pause
- Remove SystemClock from mofa-kernel pub use exports; wire up pub mod clock
  and pub use clock::SystemClock in mofa-foundation scheduler module
- Move SystemClock unit tests to mofa-foundation/src/scheduler/clock.rs
